Proof of concept NetExtender client
login.py: log in and get settings.
tunnel.py: Make a TLS connection and forward data to pppd.

Use it like this:

Log in and get a nice cookie:
$ ./login.py <server> -u <user> -p '<password>' -d <domain> --noverify

This script prints an authorization cookie if it's successful. This
cookie should be given to the next script (probably promptly):

$ sudo pppd debug nodetach file /etc/ppp/peers/sslvpn pty "./tunnel.py <server> --noverify --auth <auth cookie from login.py>"

ppp0 should now come up, whereupon you may add a route into the VPN:

$ sudo ip route add 10.0.0.0/8 dev ppp0

Known issues:
- Too much debug logging
- Routes not set up automatically
- MTU might be wrong?
- We don't handle the case when multiple frames are received in one
call to recv() on the SSL socket, or when it doesn't return a full
frame. In that case we lose sync and drop packets and screw up
royally. I believe this makes most real traffic fail at the moment.