Setup routes when ppp0 is created
The routes must be set after pppd has brought the interface up, this
usually happens in an ip-up script on the system, but in an attempt to
be less intrusive on the system a hack is created.

This will launch a thread that will loop for 3 seconds or until the ppp0
interface comes up.

If the interface comes up, it will add all the routes supplied by the
server.