Handle BlockingIOError when flushing stdout
The sys.stdout.flush() would sometime raise an exception that it would
block. Move flushing outside the function and flush at the beginning so
that any pending data is written regardless.

There is probably some corner cases if this happens at the moment when
traffic stops. The select call would probably hang and some data might
get stuck in the stdout buffer.

Anyway, this is an improvement regardless.