Make schmextender also work on Mac OS X
This change adds a tiny library called libip. By calling the function
get_impl() an object is returned which includes functions to add
addresses and routes.

Two implementations are available. One for Linux that uses the ip
utility as before. The other one is for Mac OS X and it uses the
tools route and ifconfig to perform the needed operations.

This has been tested on OS X Yosemite version 10.10.5 but should
probably work on more versions.